[ovsp4rt] Separate C API functions from rest of code
- Moved the C API wrapper functions to a separate source
  file, to allow a substitute front end to be used for
  testing.
